feat(index.ts): typescript
